requirements.txt
fastapi
uvicorn
python-dotenv
streamlit
langchain-google-genai
openai-whisper
pytubefix
qdrant-client
sentence-transformers
langchain-text-splitters
langchain-core
langchain-community
pydantic
python-multipart
